+++ src/main/java/com/vt/valuetogether/domain/user/entity/Role.java
@@ -6,10 +6,10 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum Role {
-    ROLE_ADMIN("ROLE_ADMIN", "관리자"),
-    ROLE_USER("ROLE_USER", "사용자");
+    ROLE_ADMIN("ADMIN", "관리자"),
+    ROLE_USER("USER", "사용자");
 
-    private final String code;
+    private final String authority;
 
     private final String value;
 }
